Reference

Contents

Reference

Contents

Index

Index

HFDisorderHubbard.CubicLattice

HFDisorderHubbard.HubbardPara

HFDisorderHubbard.UnitaryDecomp

HFDisorderHubbard.CubicLattice — Type

CubicLattice(L::Int)

CubicLattice with PBC boundary

HFDisorderHubbard.HubbardPara — Type

3D Anderson-Hubbard Model. Hamiltonian see index part.

HubbardPara

t : hopping amplitude

U : Hubbard interaction

W : disorder strength, onsite disorder energies are drawn from a Gaussian distribution in the interval [-W/2, W/2]

n_up : ⟨n_{i↑}⟩ = ∑_{α} |⟨i↑|α⟩|^2

n_down : ⟨n_{i↓}⟩ = ∑_{α} |⟨i↓|α ⟩|^2 [1]

S_up : ⟨S_{i}^+⟩ = ∑_{α} |⟨0|c_{i↓}^† c_{i↑}⟩|α⟩|^2

S_down : conjugation of S_up

[1] F.Fazileh et al. 2006 Physical Review Letters

HFDisorderHubbard.UnitaryDecomp — Method

UnitaryDecomp(lat::CubicLattice, para::HubbardPara)

Use Schur Decomposition to get H = U T U^† , where T is generally T is upper triangular but as H is Hermitian, T is diagonal

We only need orbitals under Fermion Surface, here output U of size (2N, N). Each row of U is a transposed conjugated eigenvector of H.